requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
motor>=3.1.1
pydantic>=1.10.7
python-multipart>=0.0.6
python-dotenv>=1.0.0
PyJWT>=2.6.0
face-recognition>=1.3.0
numpy>=1.24.2
Pillow>=9.5.0